requirements.txt
pandas>=1.3.0
psycopg2-binary>=2.9.0
matplotlib>=3.4.0
seaborn>=0.11.0
python-dotenv>=0.19.0
xlsxwriter>=3.0.0 